--- a/core/src/main/java/com/google/errorprone/bugpatterns/InterruptedExceptionSwallowed.java
+++ b/core/src/main/java/com/google/errorprone/bugpatterns/InterruptedExceptionSwallowed.java
@@ -1,0 +1,250 @@
+/*
+ * Copyright 2019 The Error Prone Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.errorprone.bugpatterns;
+
+import static com.google.errorprone.BugPattern.ProvidesFix.REQUIRES_HUMAN_ATTENTION;
+import static com.google.errorprone.BugPattern.SeverityLevel.WARNING;
+import static com.google.errorprone.matchers.Description.NO_MATCH;
+import static com.google.errorprone.util.ASTHelpers.getSymbol;
+import static com.google.errorprone.util.ASTHelpers.getType;
+import static com.google.errorprone.util.ASTHelpers.isSubtype;
+import static java.lang.Boolean.TRUE;
+
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
+import com.google.errorprone.BugPattern;
+import com.google.errorprone.VisitorState;
+import com.google.errorprone.bugpatterns.BugChecker.TryTreeMatcher;
+import com.google.errorprone.fixes.SuggestedFix;
+import com.google.errorprone.matchers.Description;
+import com.sun.source.tree.BlockTree;
+import com.sun.source.tree.CatchTree;
+import com.sun.source.tree.ClassTree;
+import com.sun.source.tree.InstanceOfTree;
+import com.sun.source.tree.LambdaExpressionTree;
+import com.sun.source.tree.MethodInvocationTree;
+import com.sun.source.tree.MethodTree;
+import com.sun.source.tree.NewClassTree;
+import com.sun.source.tree.StatementTree;
+import com.sun.source.tree.ThrowTree;
+import com.sun.source.tree.Tree;
+import com.sun.source.tree.TryTree;
+import com.sun.source.tree.VariableTree;
+import com.sun.source.util.TreeScanner;
+import com.sun.tools.javac.code.Symbol;
+import com.sun.tools.javac.code.Symbol.ClassSymbol;
+import com.sun.tools.javac.code.Symbol.MethodSymbol;
+import com.sun.tools.javac.code.Type;
+import com.sun.tools.javac.code.Type.UnionClassType;
+import com.sun.tools.javac.code.Types;
+import java.util.ArrayDeque;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+import javax.annotation.Nullable;
+
+/**
+ * Checks for cases where an {@link InterruptedException} is caught as part of a catch block
+ * catching a supertype, and not specially handled.
+ */
+@BugPattern(
+    name = "InterruptedExceptionSwallowed",
+    summary =
+        "This catch block catches a supertype of InterruptedException and doesn't handle it"
+            + " separately.",
+    severity = WARNING,
+    providesFix = REQUIRES_HUMAN_ATTENTION,
+    documentSuppression = false)
+public final class InterruptedExceptionSwallowed extends BugChecker implements TryTreeMatcher {
+  private static final String AUTOCLOSEABLE = "java.lang.AutoCloseable";
+
+  @Override
+  public Description matchTry(TryTree tree, VisitorState state) {
+    for (CatchTree catchTree : tree.getCatches()) {
+      Type type = getType(catchTree.getParameter());
+      Type interrupted = state.getSymtab().interruptedExceptionType;
+      ImmutableList<Type> caughtTypes = extractTypes(type);
+      if (caughtTypes.stream().anyMatch(t -> isSubtype(t, interrupted, state))) {
+        return NO_MATCH;
+      }
+      if (caughtTypes.stream().anyMatch(t -> isSubtype(interrupted, t, state))) {
+        ImmutableSet<Type> thrownExceptions = getThrownExceptions(tree.getBlock(), state);
+        if (thrownExceptions.stream().anyMatch(t -> isSubtype(t, interrupted, state))
+            && !blockChecksForInterruptedException(catchTree.getBlock(), state)
+            && !isSuppressed(catchTree.getParameter())) {
+          return describeMatch(catchTree, createFix(catchTree));
+        }
+      }
+    }
+    return NO_MATCH;
+  }
+
+  private static SuggestedFix createFix(CatchTree catchTree) {
+    List<? extends StatementTree> block = catchTree.getBlock().getStatements();
+    String fix =
+        String.format(
+            "if (%s instanceof InterruptedException) {\nThread.currentThread().interrupt();\n}\n",
+            catchTree.getParameter().getName());
+    if (block.isEmpty()) {
+      return SuggestedFix.replace(catchTree.getBlock(), String.format("{%s}", fix));
+    }
+    return SuggestedFix.prefixWith(block.get(0), fix);
+  }
+
+  private static boolean blockChecksForInterruptedException(BlockTree block, VisitorState state) {
+    return TRUE.equals(
+        new TreeScanner<Boolean, Void>() {
+          @Override
+          public Boolean reduce(Boolean a, Boolean b) {
+            return TRUE.equals(a) || TRUE.equals(b);
+          }
+
+          @Override
+          public Boolean visitInstanceOf(InstanceOfTree instanceOfTree, Void aVoid) {
+            return isSubtype(
+                getType(instanceOfTree.getType()),
+                state.getSymtab().interruptedExceptionType,
+                state);
+          }
+        }.scan(block, null));
+  }
+
+  private static ImmutableSet<Type> getThrownExceptions(Tree tree, VisitorState state) {
+    ScanThrownTypes scanner = new ScanThrownTypes(state);
+    scanner.scan(tree, null);
+    return ImmutableSet.copyOf(scanner.getThrownTypes());
+  }
+
+  static final class ScanThrownTypes extends TreeScanner<Void, Void> {
+    boolean inResources = false;
+    ArrayDeque<Set<Type>> thrownTypes = new ArrayDeque<>();
+
+    private final VisitorState state;
+    private final Types types;
+
+    ScanThrownTypes(VisitorState state) {
+      this.state = state;
+      this.types = state.getTypes();
+      thrownTypes.push(new HashSet<>());
+    }
+
+    private Set<Type> getThrownTypes() {
+      return thrownTypes.peek();
+    }
+
+    @Override
+    public Void visitMethodInvocation(MethodInvocationTree invocation, Void unused) {
+      MethodSymbol symbol = getSymbol(invocation);
+      if (symbol != null) {
+        getThrownTypes().addAll(symbol.getThrownTypes());
+      }
+      return super.visitMethodInvocation(invocation, null);
+    }
+
+    @Override
+    public Void visitTry(TryTree tree, Void unused) {
+      thrownTypes.push(new HashSet<>());
+      inResources = true;
+      scan(tree.getResources(), null);
+      inResources = false;
+      scan(tree.getBlock(), null);
+      scan(tree.getCatches(), null);
+      scan(tree.getFinallyBlock(), null);
+      Set<Type> fromBlock = thrownTypes.pop();
+      getThrownTypes().addAll(fromBlock);
+      return null;
+    }
+
+    @Override
+    public Void visitCatch(CatchTree tree, Void unused) {
+      Type type = getType(tree.getParameter());
+      // Remove the thrown types; if they're re-thrown they'll get re-added.
+      for (Type unionMember : extractTypes(type)) {
+        getThrownTypes().removeIf(thrownType -> types.isAssignable(thrownType, unionMember));
+      }
+      return super.visitCatch(tree, null);
+    }
+
+    @Override
+    public Void visitThrow(ThrowTree tree, Void unused) {
+      getThrownTypes().addAll(extractTypes(getType(tree.getExpression())));
+      return super.visitThrow(tree, null);
+    }
+
+    @Override
+    public Void visitNewClass(NewClassTree tree, Void unused) {
+      MethodSymbol symbol = getSymbol(tree);
+      if (symbol != null) {
+        getThrownTypes().addAll(symbol.getThrownTypes());
+      }
+      return super.visitNewClass(tree, null);
+    }
+
+    @Override
+    public Void visitVariable(VariableTree tree, Void unused) {
+      if (inResources) {
+        Symbol symbol = getSymbol(tree.getType());
+        if (symbol instanceof ClassSymbol) {
+          ClassSymbol classSymbol = (ClassSymbol) symbol;
+          Optional<MethodSymbol> methodSymbol = getCloseMethod(classSymbol, state);
+          methodSymbol.map(MethodSymbol::getThrownTypes).ifPresent(getThrownTypes()::addAll);
+        }
+      }
+      return super.visitVariable(tree, null);
+    }
+
+    // We don't need to account for anything thrown by declarations,
+    @Override
+    public Void visitLambdaExpression(LambdaExpressionTree tree, Void unused) {
+      return null;
+    }
+
+    @Override
+    public Void visitClass(ClassTree tree, Void unused) {
+      return null;
+    }
+
+    @Override
+    public Void visitMethod(MethodTree tree, Void unused) {
+      return null;
+    }
+  }
+
+  private static Optional<MethodSymbol> getCloseMethod(ClassSymbol symbol, VisitorState state) {
+    Types types = state.getTypes();
+    return symbol.getEnclosedElements().stream()
+        .filter(
+            sym ->
+                types.isAssignable(symbol.type, state.getTypeFromString(AUTOCLOSEABLE))
+                    && sym.getSimpleName().contentEquals("close")
+                    && sym.getTypeParameters().isEmpty())
+        .map(e -> (MethodSymbol) e)
+        .findFirst();
+  }
+
+  private static ImmutableList<Type> extractTypes(@Nullable Type type) {
+    if (type == null) {
+      return ImmutableList.of();
+    }
+    if (type.isUnion()) {
+      UnionClassType unionType = (UnionClassType) type;
+      return ImmutableList.copyOf(unionType.getAlternativeTypes());
+    }
+    return ImmutableList.of(type);
+  }
+}