This code has been reviewed and submitted internally. Feel free to discuss on the PR and we can submit follow-up changes as necessary.

Commits:
=====
<p> AndroidJdkLibsChecker performance improvements

This CL optimizes AndroidJdkLibsChecker#deriveApiDiff. The main improvement is that it replaces TreeSet, O(NlogN) creation, with a HashSet, O(N) creation.

                               |  before  |  after  |  diff
-------------------------------+----------+---------+----------
AndroidJdkLibsChecker.<init>   |  1.08s   |  0.59s  |  -45%
-------------------------------+----------+---------+----------

1660324cce9869eae0d0160d1c031f4e27771e6e

-------

<p> Disable ThrowSpecificExceptions in test-only code.

5496d9c9233f23dc9b3a4d3ea7843177f123b265

-------

<p> Make AbstractReturnValueIgnored visible from custom errorprone checkers, and a little more configurable:

- description can be changed
- whether return values can be ignored from exception-throwing tests is configurable

549c55ba02b6d8d0af06dce5fb0c0ec8e65f60d2

-------

<p> Let addMembers add to the beginning or end of a class

b066ed7b1d550bc0976bfe479fbf6cdd2f5e9b5b

-------

<p> Don't fire PreferJavaTimeOverload on AssertJ's assertThat() overloads.

Fixes https://github.com/google/error-prone/issues/1377

8552f69fa3e6d07b1a125353d39cb178b6ccdf61

-------

<p> InterruptedExceptionSwallowed: flag where InterruptedException is caught as part of (e.g.) Exception, and not handled.

f77b29cc42358c6084d3b4ea7513662239e61812